+++ apps/api/src/app/account-balance/account-balance.controller.ts
@@ -1,3 +1,4 @@
+import { AccountService } from '@ghostfolio/api/app/account/account.service';
 import { HasPermission } from '@ghostfolio/api/decorators/has-permission.decorator';
 import { HasPermissionGuard } from '@ghostfolio/api/guards/has-permission.guard';
 import { permissions } from '@ghostfolio/common/permissions';
@@ -5,6 +6,8 @@ import type { RequestWithUser } from '@ghostfolio/common/types';
 
 import {
   Controller,
+  Body,
+  Post,
   Delete,
   HttpException,
   Inject,
@@ -17,13 +20,49 @@ import { AccountBalance } from '@prisma/client';
 import { StatusCodes, getReasonPhrase } from 'http-status-codes';
 
 import { AccountBalanceService } from './account-balance.service';
+import { CreateAccountBalanceDto } from './create-account-balance.dto';
 
 @Controller('account-balance')
 export class AccountBalanceController {
   public constructor(
     private readonly accountBalanceService: AccountBalanceService,
+    private readonly accountService: AccountService,
     @Inject(REQUEST) private readonly request: RequestWithUser
   ) {}
+
+  @HasPermission(permissions.createAccountBalance)
+  @Post()
+  @UseGuards(AuthGuard('jwt'), HasPermissionGuard)
+  public async createAccountBalance(
+    @Body() data: CreateAccountBalanceDto
+  ): Promise<AccountBalance> {
+    const account = await this.accountService.account({
+      id_userId: {
+        id: data.accountId,
+        userId: this.request.user.id
+      }
+    });
+
+    if (!account) {
+      throw new HttpException(
+        getReasonPhrase(StatusCodes.FORBIDDEN),
+        StatusCodes.FORBIDDEN
+      );
+    }
+
+    return this.accountBalanceService.createAccountBalance({
+      Account: {
+        connect: {
+          id_userId: {
+            id: account.id,
+            userId: account.userId
+          }
+        }
+      },
+      date: data.date,
+      value: data.balance
+    });
+  }
 
   @HasPermission(permissions.deleteAccountBalance)
   @Delete(':id')