+++ libs/ui/src/lib/account-balances/account-balances.component.ts
@@ -14,7 +14,17 @@ import {
   Output,
   ViewChild
 } from '@angular/core';
+import {
+  FormGroup,
+  FormControl,
+  Validators,
+  ReactiveFormsModule
+} from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
+import { DateAdapter } from '@angular/material/core';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatSort, MatSortModule } from '@angular/material/sort';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
@@ -29,9 +39,13 @@ import { GfValueComponent } from '../value';
     CommonModule,
     GfValueComponent,
     MatButtonModule,
+    MatDatepickerModule,
+    MatFormFieldModule,
+    MatInputModule,
     MatMenuModule,
     MatSortModule,
-    MatTableModule
+    MatTableModule,
+    ReactiveFormsModule
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
   selector: 'gf-account-balances',
@@ -43,24 +57,38 @@ export class GfAccountBalancesComponent
   implements OnChanges, OnDestroy, OnInit
 {
   @Input() accountBalances: AccountBalancesResponse['balances'];
+  @Input() accountCurrency: string;
   @Input() accountId: string;
   @Input() locale = getLocale();
   @Input() showActions = true;
 
+  @Output() accountBalanceCreated = new EventEmitter<{
+    balance: number;
+    date: Date;
+  }>();
   @Output() accountBalanceDeleted = new EventEmitter<string>();
 
   @ViewChild(MatSort) sort: MatSort;
 
+  public accountBalanceForm = new FormGroup({
+    balance: new FormControl(0, Validators.required),
+    date: new FormControl(new Date(), Validators.required)
+  });
+
   public dataSource: MatTableDataSource<
     AccountBalancesResponse['balances'][0]
   > = new MatTableDataSource();
+
   public displayedColumns: string[] = ['date', 'value', 'actions'];
+  public Validators = Validators;
 
   private unsubscribeSubject = new Subject<void>();
 
-  public constructor() {}
+  public constructor(private dateAdapter: DateAdapter<any>) {}
 
-  public ngOnInit() {}
+  public ngOnInit() {
+    this.dateAdapter.setLocale(this.locale);
+  }
 
   public ngOnChanges() {
     if (this.accountBalances) {
@@ -79,6 +107,10 @@ export class GfAccountBalancesComponent
     if (confirmation) {
       this.accountBalanceDeleted.emit(aId);
     }
+  }
+
+  public onSubmitAccountBalance() {
+    this.accountBalanceCreated.emit(this.accountBalanceForm.getRawValue());
   }
 
   public ngOnDestroy() {